+++ CHANGELOG.md
@@ -9,7 +9,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Rename files separately
-- Rename movies
+
+### Fixed
+
+- Error when `config` command runned without arguments
+
+## [0.0.2] - 2021-08-03
+
+### Added
+
+- Rename movies using [TMDb API](https://www.themoviedb.org/documentation/api)
+- Choise of alternative movies with same title
+
+### Changed
+
+- CLI formatting for better user experience
+- Code refactoring
 
 ## [0.0.1] - 2021-08-02